modified:   app/Http/Controllers/BlogController.php 	modified:   app/Http/Controllers/ProfileController.php 	modified:   public/css/style.css 	new file:   public/img/icon.png 	deleted:    public/imgblog/20220909084400.jfif 	deleted:    public/imgblog/20220910014452.jpg 	deleted:    public/imgblog/20220910021244.jpg 	deleted:    public/imgblog/20220910021358.jpg 	deleted:    public/imgblog/20220910021649.jpg 	renamed:    public/imgblog/20220909083614.jpg -> public/imgblog/20220912081105.jpg 	modified:   resources/views/home.blade.php 	modified:   resources/views/include/head.blade.php 	modified:   resources/views/pages/about.blade.php 	modified:   resources/views/pages/blog.blade.php 	modified:   resources/views/pages/blogdetail.blade.php 	modified:   resources/views/pages/contact.blade.php 	modified:   resources/views/pages/services.blade.php

+++ app/Http/Controllers/BlogController.php
@@ -93,13 +93,23 @@ class BlogController extends Controller
 
         $input = $request->all();
 
-        if ($gambar = $request->file('gambar')) {
+        if($request->hasFile('gambar')){
+
+            // user intends to replace the current image for the category.  
+            // delete existing (if set)
+        
+            if($oldImage = $blog->gambar) {
+        
+                unlink(public_path('imgblog/') . $oldImage);
+            
+            }
+
+            $gambar = $request->file('gambar');
             $destinationPath = 'imgblog/';
             $profileImage = date('YmdHis') . "." . $gambar->getClientOriginalExtension();
             $gambar->move($destinationPath, $profileImage);
             $input['gambar'] = "$profileImage";
-        } else {
-            unset($input['gambar']);
+            
         }
 
         $blog->update($input);
@@ -114,7 +124,9 @@ class BlogController extends Controller
     {
         $blog = Blog::findOrFail($id);
         $image_path = public_path('imgblog').'/'.$blog->gambar;
-        unlink($image_path);
+        if(is_file($image_path)){
+            unlink($image_path);
+        }
         $blog->delete();
